Refactor saving of output
