+++ src/dials/util/combine_experiments.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
+import logging
+import random
 import sys
 from collections.abc import Sequence
-from typing import Iterator, Tuple, TypeVar
+from dataclasses import dataclass
+from typing import Iterator, Optional, Tuple, TypeVar
 
 import dxtbx.model
 import dxtbx.model.compare as compare
@@ -17,8 +20,16 @@ from libtbx import phil
 from scitbx import matrix
 
 from dials.algorithms.clustering.unit_cell import cluster_unit_cells
+from dials.algorithms.integration.stills_significance_filter import SignificanceFilter
+from dials.algorithms.integration.stills_significance_filter import (
+    phil_scope as sig_filter_phil_scope,
+)
+from dials.array_family import flex
+from dials.util import tabulate
 
 T = TypeVar("T")
+
+logger = logging.getLogger(__name__)
 
 
 def _split_equal_parts_of_length(a: Sequence[T], n: int) -> Iterator[Sequence[T]]:
@@ -295,3 +306,246 @@ def parse_ref_models(
             reference_from_experiment.average_hierarchy_level,
         )
     return (ref_beam, ref_goniometer, ref_scan, ref_crystal, ref_detector)
+
+
+@dataclass
+class significance_params:
+    significance_filter: phil.scope_extract
+
+
+def select_subset(
+    experiments: ExperimentList,
+    reflections: flex.reflection_table,
+    n_subset: int = 1,
+    n_subset_method: str = "random",
+    n_refl_panel_list=None,
+    significance_filter: Optional[phil.scope_extract] = None,
+) -> Tuple[ExperimentList, flex.reflection_table]:
+    subset_exp = ExperimentList()
+    subset_refls = flex.reflection_table()
+    if n_subset_method == "random":  # Doesn't require reflections
+        n_picked = 0
+        indices = list(range(len(experiments)))
+        if reflections.experiment_identifiers().keys():
+            indices_to_sel = []
+            while n_picked < n_subset:
+                idx = indices.pop(random.randint(0, len(indices) - 1))
+                indices_to_sel.append(idx)
+                n_picked += 1
+            # make sure select in order.
+            for idx in sorted(indices_to_sel):
+                subset_exp.append(experiments[idx])
+            subset_refls = reflections.select(subset_exp)
+            subset_refls.reset_ids()
+        else:
+            while n_picked < n_subset:
+                idx = indices.pop(random.randint(0, len(indices) - 1))
+                subset_exp.append(experiments[idx])
+                refls = reflections.select(reflections["id"] == idx)
+                refls["id"] = flex.int(len(refls), n_picked)
+                subset_refls.extend(refls)
+                n_picked += 1
+        logger.info(
+            f"Selecting a random subset of {n_subset} experiments out of {len(experiments)} total."
+        )
+    elif n_subset_method == "n_refl":  # Does require reflections
+        if n_refl_panel_list is None:
+            refls_subset = reflections
+        else:
+            sel = flex.bool(len(reflections), False)
+            for p in n_refl_panel_list:
+                sel |= reflections["panel"] == p
+            refls_subset = reflections.select(sel)
+        refl_counts = flex.int()
+        for expt_id in range(len(experiments)):
+            refl_counts.append((refls_subset["id"] == expt_id).count(True))
+        sort_order = flex.sort_permutation(refl_counts, reverse=True)
+        if reflections.experiment_identifiers().keys():
+            for idx in sorted(sort_order[:n_subset]):
+                subset_exp.append(experiments[idx])
+            subset_refls = reflections.select(subset_exp)
+            subset_refls.reset_ids()
+        else:
+            for expt_id, idx in enumerate(sort_order[:n_subset]):
+                subset_exp.append(experiments[idx])
+                refls = reflections.select(reflections["id"] == idx)
+                refls["id"] = flex.int(len(refls), expt_id)
+                subset_refls.extend(refls)
+        logger.info(
+            f"Selecting a subset of {n_subset} experiments with highest number of reflections out of {len(experiments)} total."
+        )
+
+    elif n_subset_method == "significance_filter":  # Does require reflections
+        if significance_filter is None:
+            significance_filter = sig_filter_phil_scope.extract()
+        significance_filter.enable = True
+        sig_filter = SignificanceFilter(significance_params(significance_filter))
+        refls_subset = sig_filter(experiments, reflections)
+        refl_counts = flex.int()
+        for expt_id in range(len(experiments)):
+            refl_counts.append((refls_subset["id"] == expt_id).count(True))
+        sort_order = flex.sort_permutation(refl_counts, reverse=True)
+        if reflections.experiment_identifiers().keys():
+            for idx in sorted(sort_order[:n_subset]):
+                subset_exp.append(experiments[idx])
+            subset_refls = reflections.select(subset_exp)
+            subset_refls.reset_ids()
+        else:
+            for expt_id, idx in enumerate(sort_order[:n_subset]):
+                subset_exp.append(experiments[idx])
+                refls = reflections.select(reflections["id"] == idx)
+                refls["id"] = flex.int(len(refls), expt_id)
+                subset_refls.extend(refls)
+    else:
+        raise ValueError(f"Invalid option for n_subset_method: {n_subset_method}")
+
+    return subset_exp, subset_refls
+
+
+def combine_experiments(params, experiment_lists, reflection_tables):
+    """Run combine_experiments, but allow passing in of parameters"""
+
+    flat_exps = ExperimentList()
+    try:
+        for elist in experiment_lists:
+            flat_exps.extend(elist)
+    except RuntimeError:
+        sys.exit(  # FIXME
+            "Unable to combine experiments. Are experiment IDs unique? "
+            "You may need to run dials.assign_experiment_identifiers first to "
+            "reset IDs."
+        )
+
+    ref_beam, ref_goniometer, ref_scan, ref_crystal, ref_detector = parse_ref_models(
+        flat_exps, params.reference_from_experiment
+    )
+
+    combine = CombineWithReference(
+        beam=ref_beam,
+        goniometer=ref_goniometer,
+        scan=ref_scan,
+        crystal=ref_crystal,
+        detector=ref_detector,
+        params=params,
+    )
+
+    # set up global experiments and reflections lists
+    reflections = flex.reflection_table()
+    experiments = ExperimentList()
+    global_id = 0
+    skipped_expts_min_refl = 0
+    skipped_expts_max_refl = 0
+
+    # loop through the input, building up the global lists
+    nrefs_per_exp = []
+    for refs, exps in zip(reflection_tables, experiment_lists):
+        # Record initial mapping of ids for updating later.
+        ids_map = dict(refs.experiment_identifiers())
+        # Keep track of mapping of imageset_ids old->new within this experimentlist
+        imageset_result_map = {}
+
+        for k in refs.experiment_identifiers().keys():
+            del refs.experiment_identifiers()[k]
+        for i, exp in enumerate(exps):
+            sel = refs["id"] == i
+            sub_ref = refs.select(sel)
+            n_sub_ref = len(sub_ref)
+            if (
+                params.output.min_reflections_per_experiment is not None
+                and n_sub_ref < params.output.min_reflections_per_experiment
+            ):
+                skipped_expts_min_refl += 1
+                continue
+            if (
+                params.output.max_reflections_per_experiment is not None
+                and n_sub_ref > params.output.max_reflections_per_experiment
+            ):
+                skipped_expts_max_refl += 1
+                continue
+
+            nrefs_per_exp.append(n_sub_ref)
+            sub_ref["id"] = flex.int(len(sub_ref), global_id)
+
+            # now update identifiers if set.
+            if i in ids_map:
+                sub_ref.experiment_identifiers()[global_id] = ids_map[i]
+            if params.output.delete_shoeboxes and "shoebox" in sub_ref:
+                del sub_ref["shoebox"]
+
+            try:
+                experiments.append(combine(exp))
+            except ComparisonError as e:
+                # When we failed tolerance checks, give a useful error message
+                (path, index) = find_experiment_in(exp, params.input.experiments)
+                sys.exit(  # FIXME - raise RuntimeError?
+                    "Model didn't match reference within required tolerance for experiment {} in {}:"
+                    "\n{}\nAdjust tolerances or set compare_models=False to ignore differences.".format(
+                        index, path, str(e)
+                    )
+                )
+
+            # Rewrite imageset_id, if the experiment has an imageset
+            if exp.imageset and "imageset_id" in sub_ref:
+                # Get the index of the imageset for this experiment and record how it changed
+                new_imageset_id = experiments.imagesets().index(
+                    experiments[-1].imageset
+                )
+                old_imageset_id = exps.imagesets().index(exp.imageset)
+                imageset_result_map[old_imageset_id] = new_imageset_id
+
+                # Check for invalid(?) imageset_id indices... and leave if they are wrong
+                if len(set(sub_ref["imageset_id"])) != 1:
+                    logger.warning(
+                        "Warning: Experiment %d reflections appear to have come from multiple imagesets - output may be incorrect",
+                        i,
+                    )
+                else:
+                    sub_ref["imageset_id"] = flex.int(len(sub_ref), new_imageset_id)
+
+            reflections.extend(sub_ref)
+
+            global_id += 1
+
+        # Include unindexed reflections, if we can safely remap their imagesets
+        if "imageset_id" in reflections:
+            unindexed_refs = refs.select(refs["id"] == -1)
+            for old_id in set(unindexed_refs["imageset_id"]):
+                subs = unindexed_refs.select(unindexed_refs["imageset_id"] == old_id)
+                subs["imageset_id"] = flex.int(len(subs), imageset_result_map[old_id])
+                reflections.extend(subs)
+
+    # Finished building global lists
+
+    if (
+        params.output.min_reflections_per_experiment is not None
+        and skipped_expts_min_refl > 0
+    ):
+        logger.info(
+            f"Removed {skipped_expts_min_refl} experiments with fewer than {params.output.min_reflections_per_experiment} reflections"
+        )
+    if (
+        params.output.max_reflections_per_experiment is not None
+        and skipped_expts_max_refl > 0
+    ):
+        logger.info(
+            f"Removed {skipped_expts_max_refl} experiments with more than {params.output.max_reflections_per_experiment} reflections"
+        )
+
+    # print number of reflections per experiment
+
+    header = ["Experiment", "Number of reflections"]
+    rows = [(str(i), str(n)) for (i, n) in enumerate(nrefs_per_exp)]
+    logger.info(tabulate(rows, header))
+
+    # select a subset if requested
+    if params.output.n_subset is not None and len(experiments) > params.output.n_subset:
+        experiments, reflections = select_subset(
+            experiments,
+            reflections,
+            n_subset=params.output.n_subset,
+            n_subset_method=params.output.n_subset_method,
+            n_refl_panel_list=params.output.n_refl_panel_list,
+            significance_filter=params.output.significance_filter,
+        )
+
+    return experiments, reflections
